♻️refactor: SupabaseManager가 AppSecrets를 사용하도록 구조 개선
